Move deref closure builder to `clippy_utils::sugg` module
